requirements.txt
Django==3.2.*
Pillow==8.3.*
django-debug-toolbar==3.2.*
django-filter==2.4.*
django-htmx==1.2.*
django-import-export==2.5.*
django-cleanup==5.2.*
django-compressor==2.4.*
django-mptt==0.13.*
django-mptt-admin==2.1.*
WeasyPrint==51
django-tinymce==3.3.*
sentry-sdk==1.4.*
psycopg2==2.9.*
psycopg2-binary==2.9.*
redis==3.5.*
gunicorn==20.1.*
whitenoise==5.3.*
celery==4.4.*
django-colorfield==0.4.*
django-solo==1.2.*